refactor: update import path for mySocialLinks and remove export from utils
