+++ src/utils/index.ts
@@ -1,3 +1,1 @@
-export * from './social-networks'
 export * from './utils'
-
